Add Functor instance for Interval.
I believe this was originally present in sivertb's private repository,
but didn't get pushed to GitHub before his mysterious disappearance.
Another project of him, called woburn, requires that Interval has a
Functor instance, as I wanted bad to build it, I had to make a guess at
what this instance is supposed to do. There are 3 choices to it: either
apply function to the start or the end of the interval, or to both. The
case at hand suggests the latter to be the correct version, as it's used
to scale the whole Diet object, which would require applying function to
both ends of every interval -- otherwise the intervals may overlap.